Detecting recipes that should be filtered. Using map and filter very poorly
